Enable Qt 5 connection syntax for QAbstractButton::animateClick()
Slot QAbstractButton::animateClick() is commonly connected to
signal QLineEdit::returnPressed() in dialogs.

The function had a parameter with a default for specifying the delay
which requires a lambda when using Qt 5 connection syntax.

Since the default value is used practically everywhere, remove
the parameter.

[ChangeLog][QtWidgets][QAbstractButton] The interval parameter has been
removed from QAbstractButton::animateClick().

Task-number: QTBUG-81845
Change-Id: I4e399138b6aa3a9f0b31c1f10de97ffcbab6a337
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>