Fix QTextDocumentLayout test with certain fonts
If there is a negative right bearing on the last character of the string,
it will stretch past the tab position (which aligns the advance position,
and not the right edge of the glyph). For certain fonts, this would cause
the actual ideal width to be calculated as 301 instead of 300 (which is
the right alignment edge set in the code).

In 5.15, this test now also passes on WinRT, so remove the QEXPECT_FAIL.

Fixes: QTBUG-46206
Task-number: QTBUG-68297
Change-Id: I03e8b8fb86e9ebe5337d3ba3384ade73d2ccdd69
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Lars Knoll <lars.knoll@qt.io>
(cherry picked from commit 8231614661e1eecaabcf33d1e332809556e86089)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>