QTextOdfWriter: store all bulletChar()s as char16_t's
... and return them as QStringView from a file-static function so that
the conversion to QString is centralized in just one place (and we can
think of skipping the QString conversion once QXmlStreamWriter can
write QStringViews and not just QStrings).

Because the Style enum is ... weird (negative values), plaster the
code with static_asserts so that we get to detect breakage when the
enum values change.

Change-Id: I4ca89b6c2601c6a1153e202de966356bb4f51651
Reviewed-by: Mårten Nordheim <marten.nordheim@qt.io>