Revert "QVector: add a construction from QArrayDataPointerRef"
This does not work properly: QVector::detach() does not
cope with statically allocated data, because it tries to
reallocate the capacity, which is 0 for static data (!),
completely ignoring the size and therefore not copying
any elements.

Qt 6 is affected, but we'll try to fix Q_ARRAY_LITERAL
there.

We can happily revert, it's private API anyhow, and the
only user so far has also been removed.

This reverts commit 43c410867b5eccea2e0f7eeceabad76d5f874101.

Task-number: QTBUG-84069
Change-Id: Idbace873fc0bdb7271ab7793d57ef4e0892ac406
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>