Revert "WebGradients: redo implementation"
Q_ARRAY_LITERAL does not work properly in Qt 5, and no
fix is planned for it, so this implementation (that relies on it)
must be removed as well; it's causing regressions from Qt 5.14.

More in detail: a QVector built using Q_ARRAY_LITERAL will
not copy the existing contents when it detaches. This corrupts
the preset gradients.

A followup commit will also revert the corresponding QVector
constructor so that no one will try to use it.

This reverts commit 5b4b437b30b320e2cd7c9a566999a39772e5d431.

Task-number: QTBUG-84069
Change-Id: Ia828f3e8a124331b58e0322e322707f4d54769f0
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>